updated optical properties

+++ src/FT0PrimaryGeneratorAction.cc
@@ -78,12 +78,16 @@ void FT0PrimaryGeneratorAction::GeneratePrimaries(G4Event* event)
 {
   auto particleTable = G4ParticleTable::GetParticleTable();
   fParticleGun->SetParticlePosition(G4ThreeVector(0,0,0));
+  
+  //fParticleGun->SetParticlePosition(G4ThreeVector(0,74.30*mm,3800*mm));
+  
   fParticleGun->SetParticleDefinition(particleTable->FindParticle("mu+"));
   fParticleGun->SetParticleEnergy(1.*GeV);
   
-  G4double x = (-26.50-dist/2)*mm + ((float) rand()/RAND_MAX)*(53.00+dist)*mm;
-  G4double y = (74.30-26.50-dist/2)*mm + ((float) rand()/RAND_MAX)*(53.00+dist)*mm;
+  G4double x = (-26.50-dist/2)*mm + ((double) rand()/RAND_MAX)*(53.00+dist)*mm;
+  G4double y = (74.30-26.50-dist/2)*mm + ((double) rand()/RAND_MAX)*(53.00+dist)*mm;
   G4double z = 3332.945*mm;
+  
   
   fParticleGun->SetParticleMomentumDirection(G4ThreeVector(x,y,z));
   fParticleGun->GeneratePrimaryVertex(event);
